Fixes for Dropdown and Refactoring
