+++ apps/web/styles/index.module.css
@@ -1,5 +1,6 @@
 .container {
   text-align: center;
   max-width: 1240px;
+  width: 100%;
   margin: 40px auto;
 }
