--- packages/ui/src/dropdown-menu/DropdownOption.tsx
@@ -1,22 +1,0 @@
-import { StyleSheet, TouchableOpacity } from "react-native";
-import { ReactNode } from "react";
-
-export const MenuOption = ({
-  onSelect,
-  children,
-}: {
-  onSelect: () => void;
-  children: ReactNode;
-}) => {
-  return (
-    <TouchableOpacity onPress={onSelect} style={styles.menuOption}>
-      {children}
-    </TouchableOpacity>
-  );
-};
-const styles = StyleSheet.create({
-  menuOption: {
-    backgroundColor: "green",
-    padding: 5,
-  },
-});
