+++ packages/ui/src/expenses/index.tsx
@@ -1,3 +1,2 @@
-export { ExpenseList, type ExpenseListProps } from "./expense-list";
-export { ExpenseItem, type ExpenseItemProps } from "./expense-item";
-// export { getExpenseListFromApi } from "./actions";
+export { ExpenseList, type ExpenseListProps } from "./ExpenseList";
+export { ExpenseItem, type ExpenseItemProps } from "./ExpenseItem";
